+++ src/Core/Managed/Shared/Extensibility/IMetricProcessor.cs
@@ -1,0 +1,18 @@
+﻿namespace Microsoft.ApplicationInsights.Extensibility
+{
+    using System;
+    using System.Collections.Generic;
+
+    /// <summary>
+    /// Provides functionality to process metric values prior to aggregation.
+    /// </summary>
+    public interface IMetricProcessor
+    {
+        /// <summary>
+        /// Process metric value.
+        /// </summary>
+        /// <param name="metric">Metric definition.</param>
+        /// <param name="value">Metric value.</param>
+        void Track(Metric metric, double value);
+    }
+}